Move stream::client to client::stream
