+++ src/client/conn/transport/tls.rs
@@ -6,8 +6,8 @@ use http::Uri;
 use rustls::ClientConfig as TlsClientConfig;
 
 use super::{TlsConnectionError, Transport, TransportStream};
+use crate::client::stream::Stream as ClientStream;
 use crate::info::HasConnectionInfo;
-use crate::stream::client::Stream as ClientStream;
 
 /// Transport via TLS
 #[derive(Debug, Clone)]
