+++ src/discovery/transport.rs
@@ -10,8 +10,8 @@ use hyper::Uri;
 
 use super::ConnectionError;
 use super::ServiceRegistry;
+use crate::client::stream::Stream as ClientStream;
 use crate::client::TransportStream;
-use crate::stream::client::Stream as ClientStream;
 
 pub use builder::TransportBuilder;
