+++ tests/server.rs
@@ -8,9 +8,9 @@ use http_body_util::BodyExt as _;
 use hyper::Response;
 use hyperdriver::bridge::rt::TokioExecutor;
 use hyperdriver::client::conn::Connection as _;
+use hyperdriver::client::stream::Stream;
 use hyperdriver::client::HttpProtocol;
 use hyperdriver::service::MakeServiceRef;
-use hyperdriver::stream::client::Stream;
 use hyperdriver::stream::server::Accept;
 
 use hyperdriver::server::{Protocol, Server};
